Refactor templating for more flexibility
Use a "render" python package to render the configuration and terraform configuration.
It uses the inventory (`inventory/inventory.yml`) to generate them dynamically. This way only the inventory is needed to add a standardized ixp or vlt server.

Fun fact, it is all vibe coded using Gemini 2.5 Pro.
I was particularly unmotivated to do this boring change, but it was on my todo list for too long.